+++ VisaCheckout_IOS_SDK/VisaCheckoutSDK.framework/Headers/VisaCheckout.h
@@ -10,9 +10,4 @@ FOUNDATION_EXPORT double VisaCheckoutVersionNumber;
 //! Project version string for VisaCheckout.
 FOUNDATION_EXPORT const unsigned char VisaCheckoutVersionString[];
 
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wdocumentation"
-
 #import "VisaCheckoutSDK-Swift.h"
-
-#pragma clang diagnostic pop